WordFlag -changes common flag of word (removed from WordList)
