+++ wordle-web/src/views/WordEditorView.vue
@@ -18,4 +18,5 @@
 <script setup lang="ts">
 import UpdateWordList from '../components/UpdateWordList.vue'
 import WordList from '../components/WordList.vue'
+import WordFlag from '../components/WordFlag.vue';
 </script>
